skill_agent/Python/requirements.txt
# Skill Agent Python Dependencies

# Web框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# 向量数据库
chromadb>=0.4.22

# 嵌入模型 - Qwen3需要更高版本
sentence-transformers>=2.7.0  # Qwen3最低要求
torch>=2.0.0
transformers>=4.51.0  # Qwen3必需，否则报错 KeyError: 'qwen3'

# 文件监听
watchdog>=4.0.0

# 工具库
pyyaml>=6.0.1
numpy  # 使用最新版本
pandas>=2.1.4

# HTTP客户端（用于测试）
httpx>=0.26.0
requests>=2.31.0

# 日志和缓存
python-json-logger>=2.0.7
cachetools>=5.3.2

# 可选：进度条
tqdm>=4.66.1
